error fix

+++ scripts/eb_deploy.sh
@@ -3,14 +3,12 @@
 #exit if there is an error
 set -e
 
-#if [-z "$1"]
-#then
-#    echo "usage $0 eb_env_name"
-#    exit 1
-#fi
+if [ "$#" -ne 1 ]; then
+    echo "usage $0 eb_env_name"
+    exit 1
+fi
 
-eb_env=$2
-echo ${eb_env}
+eb_env=$1
 
 rm -rf .elasticbeanstalk
 
@@ -25,7 +23,6 @@ deploy:
 _EOF
 
 #deploy
-echo "deploy"
 eb deploy
 
 exit 0
